refactor(CardView): split CarView into more meaningful files
* Add GameCardView for CardView used in Game
* CardView now deals only with the View itself, there is no logic at all
  in it
* Fix #9
