+++ Shared/Cards/VictoryCardsView.swift
@@ -12,10 +12,10 @@ struct VictoryCardsView: View {
             Spacer()
             ZStack {
                 ForEach(0..<5) { cardNumber in
-                    CCardView(
+                    CardView(
                         color: colors[cardNumber],
                         image: images[cardNumber],
-                        isFacedUp: $isCardsFacedUp[cardNumber]
+                        isFacedUp: isCardsFacedUp[cardNumber]
                     )
                     .frame(width: 80, height: 80)
                     .rotationEffect(angle(cardNumber: cardNumber))
@@ -26,24 +26,8 @@ struct VictoryCardsView: View {
             }
             Spacer()
         }
-        .onAppear {
-            returnCardTask = Timer.publish(every: 1, on: .main, in: .default)
-                .autoconnect()
-                .sink { _ in
-                    isCardsFacedUp = [Bool.random(), Bool.random(), Bool.random(), Bool.random(), Bool.random()]
-                    colors = isCardsFacedUp.enumerated().map { offset, isFacedUp in
-                        if isFacedUp { return [.green, .blue, .red].randomElement() ?? .red }
-                        return colors[offset]
-                    }
-                    images = isCardsFacedUp.enumerated().map { offset, isFacedUp in
-                        if isFacedUp { return images[offset] }
-                        return Art.allCases.randomElement()?.image ?? Art.artDeco.image
-                    }
-                }
-        }
-        .onDisappear {
-            returnCardTask?.cancel()
-        }
+        .onAppear(perform: startVictoryAnimation)
+        .onDisappear(perform: returnCardTask?.cancel)
     }
 
     private func angle(cardNumber: Int) -> Angle {
@@ -64,49 +48,21 @@ struct VictoryCardsView: View {
     private var randomArt: Image {
         Art.allCases.randomElement()?.image ?? Art.artDeco.image
     }
-}
 
-// TODO: refactor this with the one in CardView
-struct CCardView: View {
-    let color: Color
-    let image: Image
-    @Binding var isFacedUp: Bool
-    let action: () -> Void = { }
-    private static let turnCardAnimationDuration: Double = 2/5
-
-    var body: some View {
-        ZStack {
-            if !isFacedUp {
-                Button(action: action) {
-                    color.transition(turnTransition)
+    private func startVictoryAnimation() {
+        returnCardTask = Timer.publish(every: 1, on: .main, in: .default)
+            .autoconnect()
+            .sink { _ in
+                isCardsFacedUp = [Bool.random(), Bool.random(), Bool.random(), Bool.random(), Bool.random()]
+                colors = isCardsFacedUp.enumerated().map { offset, isFacedUp in
+                    if isFacedUp { return [.green, .blue, .red].randomElement() ?? .red }
+                    return colors[offset]
                 }
-                .buttonStyle(PlainButtonStyle())
-            } else {
-                image
-                    .renderingMode(.original)
-                    .resizable()
-                    .font(.largeTitle)
-                    .transition(turnTransition)
+                images = isCardsFacedUp.enumerated().map { offset, isFacedUp in
+                    if isFacedUp { return images[offset] }
+                    return Art.allCases.randomElement()?.image ?? Art.artDeco.image
+                }
             }
-        }
-        .modifier(AddCardStyle(foregroundColor: color))
-        .rotation3DEffect(
-            isFacedUp
-                ? .radians(.pi)
-                : .zero,
-            axis: (x: 0.0, y: 1.0, z: 0.0),
-            perspective: 1/3
-        )
-        .animation(.easeInOut(duration: Self.turnCardAnimationDuration))
-        .rotation3DEffect(.radians(.pi), axis: (x: 0.0, y: 1.0, z: 0.0))
-    }
-
-    private var turnTransition: AnyTransition {
-        AnyTransition.opacity.animation(
-            Animation
-                .linear(duration: 0.01)
-                .delay(Self.turnCardAnimationDuration/2)
-        )
     }
 }
