+++ iOS/MainView-iOS.swift
@@ -22,7 +22,7 @@ struct MainView: View {
                     GameOverView(store: store.gameStore)
                     adaptiveGrid(level: viewStore.game.level) {
                         ForEach(viewStore.game.cards) {
-                            CardView(store: store.gameStore, card: $0)
+                            GameCardView(store: store.gameStore, card: $0)
                         }
                     }
                     .padding()
